feat: improve wattsapp and phone link
